+++ examples/example-junit-spring-systemtest/src/test/java/com/fitbur/testify/examples/junit/spring/systemtest/resource/GreeterResourceDirectTest.java
@@ -24,7 +24,6 @@ import com.fitbur.testify.system.SpringSystemTest;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  *