+++ need/need-hsql/src/main/java/com/fitbur/testify/need/hsql/InMemoryHSQL.java
@@ -15,13 +15,9 @@
  */
 package com.fitbur.testify.need.hsql;
 
-import static com.fitbur.guava.common.base.Preconditions.checkState;
 import com.fitbur.testify.need.NeedDescriptor;
 import com.fitbur.testify.need.NeedProvider;
 import static java.lang.String.format;
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.sql.Statement;
 import org.hsqldb.jdbc.JDBCDataSource;
 
 /**
@@ -34,43 +30,12 @@ public class InMemoryHSQL implements NeedProvider<JDBCDataSource> {
     @Override
     public JDBCDataSource configuration(NeedDescriptor descriptor) {
         JDBCDataSource dataSource = new JDBCDataSource();
-        dataSource.setUrl(format("jdbc:hsqldb:mem:%s", descriptor.getTestClassName()));
+        dataSource.setUrl(format("jdbc:hsqldb:mem:%s_%s",
+                descriptor.getTestClassName(), descriptor.getTestMethodName()));
         dataSource.setUser("sa");
         dataSource.setPassword("");
 
         return dataSource;
-    }
-
-    @Override
-    public void init(NeedDescriptor descriptor, JDBCDataSource dataSource) {
-        try {
-            try (Connection connection = dataSource.getConnection()) {
-                Statement statement = connection.createStatement();
-                statement.execute("TRUNCATE SCHEMA PUBLIC RESTART IDENTITY AND COMMIT NO CHECK");
-            } catch (SQLException e) {
-                throw e;
-            }
-
-        } catch (SQLException e) {
-            checkState(false, "Need provider '%s' 'init' failed.\n%s",
-                    this.getClass().getSimpleName(), e.getMessage());
-        }
-    }
-
-    @Override
-    public void destroy(NeedDescriptor descriptor, JDBCDataSource dataSource) {
-        try {
-            try (Connection connection = dataSource.getConnection()) {
-                Statement statement = connection.createStatement();
-                statement.execute("SHUTDOWN");
-            } catch (SQLException e) {
-                throw e;
-            }
-
-        } catch (SQLException e) {
-            checkState(false, "Need provider '%' failed shutdown the database.\n%s",
-                    this.getClass().getSimpleName(), e.getMessage());
-        }
     }
 
 }